conformance/functionalityTesting/enqueueWriteBuffer.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>enqueueWriteBuffer conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>

function getRandomNumber (min, max) {
    return Math.random() * (max - min) + min;
}

var executeKernel = function (webCLContext, kernelName, webCLDevices, webCLCommandQueue, data, result) {
    try {
        input = webCLContext.createBuffer(webcl.MEM_READ_WRITE, DATA_SIZE);
        output = webCLContext.createBuffer(webcl.MEM_READ_WRITE, DATA_SIZE);
        webCLCommandQueue.enqueueWriteBuffer(input, true, 0, DATA_SIZE, data);

        kernelName = kernelName.replace(/^\s+|\s+$/g, '');
        var kernelPath = "../../resources/kernels/" + kernelName + ".cl";
        kernelSource = wtu.readKernel(kernelPath);
        webCLProgram = wtu.createProgram(webCLContext, kernelSource);
        wtu.build(webCLProgram, webCLDevices);

        webCLKernel = wtu.createKernel(webCLProgram, kernelName);
        webCLKernel.setArg(0, input);
        webCLKernel.setArg(1, output);

        globalWorkSize[0] = SIZE;
        localWorkSize[0] = webCLKernel.getWorkGroupInfo(webCLDevices[0], webcl.KERNEL_WORK_GROUP_SIZE);
        shouldBe("wtu.enqueueNDRangeKernel(webCLCommandQueue, webCLKernel, null, globalWorkSize, localWorkSize);", "true", true);
        webCLCommandQueue.enqueueReadBuffer(output, true, 0, SIZE * data.BYTES_PER_ELEMENT, result);
        correct = 0;
        for (i = 0; i < count; i++)
            if (result[i] == data[i])
                correct++;

        console.log("Computed " + correct + "/" + count + " correct values for kernel: " + kernelName);
        if (count == correct)
            testPassed("Test passed for all " + count + " values for kernel: " + kernelName);
        else
            testFailed("Test failed for kernel: " + kernelName + ". Computed " + correct + " / " + count + " correct values.");

    } catch (e) {
        testFailed("Test failed for kernel: " + kernelName + " . Exception thrown: " + e.message);
    }
}

try {
    debug("enqueueWriteBuffer functionality test case.");

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLPlatforms = wtu.getPlatforms();
    var webCLDevices = wtu.getDevices(webCLPlatforms);
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var kernelSource, webCLKernel, webCLProgram;
    var correct, input, output;
    var count = SIZE;
    var dataInt = new Int32Array(SIZE);
    var dataIntShort = new Int8Array(SIZE);
    var dataFloat = new Float32Array(SIZE);
    var resultInt = new Int32Array(SIZE);
    var resultIntShort = new Int8Array(SIZE);
    var resultFloat = new Float32Array(SIZE);
    var globalWorkSize = new Int32Array(1);
    var localWorkSize = new Int32Array(1);

    /*
    *   Testing for kernel "test_stream_write_int".
    */

    for (i = 0; i < count; i++)
        dataInt[i] = getRandomNumber(-2147483648, 2147483647);

    executeKernel(webCLContext, "test_stream_write_int", webCLDevices, webCLCommandQueue, dataInt, resultInt);

    /*
    *   Testing for kernel "test_stream_write_uint".
    */

    for (i = 0; i < count; i++)
        dataInt[i] = getRandomNumber(0, 4294967295);

    executeKernel(webCLContext, "test_stream_write_uint", webCLDevices, webCLCommandQueue, dataInt, resultInt);

    /*
    *   Testing for kernel "test_stream_write_float".
    */

    for (i = 0; i < count; i++)
        dataFloat[i] = Math.random() * 10;

    executeKernel(webCLContext, "test_stream_write_float", webCLDevices, webCLCommandQueue, dataFloat, resultFloat);

    /*
    *   Testing for kernel "test_stream_write_long".
    */

    for (i = 0; i < count; i++)
        dataFloat[i] = getRandomNumber(-9223372036854775808, 9223372036854775807);

    executeKernel(webCLContext, "test_stream_write_long", webCLDevices, webCLCommandQueue, dataFloat, resultFloat);

    /*
    *   Testing for kernel "test_stream_write_ulong".
    */

    for (i = 0; i < count; i++)
        dataFloat[i] = getRandomNumber(0, 18446744073709551615);

    executeKernel(webCLContext, "test_stream_write_ulong", webCLDevices, webCLCommandQueue, dataFloat, resultFloat);

    /*
    *   Testing for kernel "test_stream_write_short".
    */

    for (i = 0; i < count; i++)
        dataIntShort[i] = getRandomNumber(-32768, 32767);

    executeKernel(webCLContext, "test_stream_write_short", webCLDevices, webCLCommandQueue, dataIntShort, resultIntShort);

    /*
    *   Testing for kernel "test_stream_write_ushort".
    */

    for (i = 0; i < count; i++)
        dataIntShort[i] = getRandomNumber(0, 65535);

    executeKernel(webCLContext, "test_stream_write_ushort", webCLDevices, webCLCommandQueue, dataIntShort, resultIntShort);

    /*
    *   Testing for kernel "test_stream_write_char".
    */

    for (i = 0; i < count; i++)
        dataIntShort[i] = getRandomNumber(-128, 127);

    executeKernel(webCLContext, "test_stream_write_char", webCLDevices, webCLCommandQueue, dataIntShort, resultIntShort);

    /*
    *   Testing for kernel "test_stream_write_uchar".
    */

    for (i = 0; i < count; i++)
        dataIntShort[i] = getRandomNumber(0, 255);

    executeKernel(webCLContext, "test_stream_write_uchar", webCLDevices, webCLCommandQueue, dataIntShort, resultIntShort);

} catch(e) {
    testFailed("enqueueWriteBuffer throws an exception : " + e.message);
}
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
